Develop

+++ app/Http/Controllers/Web/FoodsAndDrinks/ProductController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\UpdateIngredientRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Models\Material;
 use App\Models\Product;
-use App\Models\Promotion;
 use App\Transformers\ProductTransformer;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -21,6 +20,7 @@ class ProductController extends WebBaseController
 
     public function index(Request $request)
     {
+        $this->authorize('index', Product::class);
         if ($request->ajax){
             $drinks_product = Product::where('type', 'Drink')
                 ->orderBy('created_at', 'desc')
@@ -45,17 +45,18 @@ class ProductController extends WebBaseController
     }
 
     public function show(Product $product){
+        $this->authorize('show', Product::class);
         $ingredient_orther = Material::whereDoesntHave('products',function($query) use ($product) {
             $query->where('product_id', $product->id);
         })->get();
         return response()->json([
             'product'               =>  (new ProductTransformer)->transform($product),
             'ingredient_orther'     => $ingredient_orther,
-            'promotions'            => Promotion::all()
         ], 200);
     }
 
     public function update(UpdateProductRequest $request, Product $product){
+        $this->authorize('update', Product::class);
         DB::beginTransaction();
         try {
             if ($request->url){
@@ -67,18 +68,12 @@ class ProductController extends WebBaseController
                 $request->url->move($destinationPath, $profileImage);
                 $product->url = $profileImage;
             }
-            if ($request->promotion_id) {
-                $promotion = Promotion::find($request->promotion_id);
-                if ($promotion){
-                    $product->promotion_id = $promotion->id;
-                }
-            }
-            else{
-                $product->promotion_id = null;
-            }
-            $product->fill($request->except(['url','sale_price', 'promotion_id']));
+            $product->fill($request->except(['url','sale_price']));
             $product->save();
             DB::commit();
+            return response()->json([
+                'status' => 'sucess'
+            ],200);
         }
         catch (\Exception $exception){
             DB::rollBack();
@@ -89,6 +84,7 @@ class ProductController extends WebBaseController
     }
 
     public function destroy(Product $product){
+        $this->authorize('destroy', Product::class);
         $product->materials()->detach();
         if ($product->url != 'default_url_product.png' ){
             File::delete(public_path('images\products\\' .$product->url));
@@ -100,6 +96,7 @@ class ProductController extends WebBaseController
     }
 
     public function updateIngredient(UpdateIngredientRequest $request, Product $product){
+        $this->authorize('updateIngredient', Product::class);
         $product->materials()->attach($request->material_id, [
             'quantity'  => $request->quantity,
             'unit'      => $request->unit
@@ -115,6 +112,7 @@ class ProductController extends WebBaseController
     }
 
     public function deleteIngredient(Product $product, Material $material){
+        $this->authorize('deleteIngredient', Product::class);
         $product->materials()->detach($material->id);
         $ingredient_orther = Material::whereDoesntHave('products',function($query) use ($product) {
             $query->where('product_id', $product->id);
@@ -127,10 +125,11 @@ class ProductController extends WebBaseController
     }
 
     public function store(CreateProductRequest $request){
+        $this->authorize('store', Product::class);
         DB::beginTransaction();
         try {
             $product = new  Product();
-            $product->fill( $request->only(['name', 'price', 'type']));
+            $product->fill( $request->only(['name', 'price', 'type', 'description']));
             $product = Product::query()->create($request->except('url'));
             if ($request->url){
                 $destinationPath = 'images/products/';
@@ -138,16 +137,17 @@ class ProductController extends WebBaseController
                 $request->url->move($destinationPath, $profileImage);
                 $product->url = $profileImage;
             }
-            $product->save();
+           $product->save();
+            DB::commit();
             return response()->json([
                 'status' => 'success'
             ],201);
-            DB::commit();
         }
         catch (\Exception $exception){
             DB::rollBack();
             return response()->json([
-                'status' => 'fails'
+                'status'    => 'fails',
+                'message'   => $exception->getMessage()
             ],422);
         }
